Rename `nested_dissection` module to `ordering`
