+++ src/sparse/linalg/ordering.rs
